+++ internal/controller/addon_controller.go
@@ -192,6 +192,7 @@ func (r *AddonReconciler) Reconcile(ctx context.Context, req ctrl.Request) (ctrl
 
 	if cm.IsConditionTrue(conditions.TypeSynced) && cm.IsConditionTrue(conditions.TypeHealthy) {
 		cm.SetReady(conditions.ReasonFullyReconciled, "Addon is fully reconciled and healthy")
+		addon.Status.Deployed = true
 	} else if !cm.IsDegraded() {
 		reason := conditions.ReasonWaitingForSync
 		message := "Waiting for Application to sync"